pyproject.toml
[project]
name = "mathexpressionapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "antlr4-python3-runtime==4.11",
    "fastapi>=0.129.2",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sympy>=1.14.0",
    "tensorflow>=2.20.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
